build(mypy): clear 10 validation modules (no-any-return, savez stub)
Third increment of the SYS-AUDIT-02-MYPY1 burn-down. Resolve every mypy
error in ten validation modules and remove them from the targeted ignore
list (54 modules remain, 74 now strict-checked).

- no-any-return: return the concrete declared type instead of an Any leak.
  float(...) for the scalar rel-error helpers whose expressions go Any
  through untyped physics constants (validate_current_drive
  critical_energy, validate_runaway_electron collision_time,
  validate_sol_two_point conduction_integral); an annotated local for the
  dict payloads (benchmark_capacitor_bank_state subprocess JSON,
  benchmark_vmec_lite_claims geometry report) and the ndarray return
  (disruption_channel_recipes dbdt_gauss_per_s).
- savez stub: numpy 2.2.6 types savez/savez_compressed as
  (*args: ArrayLike, allow_pickle: bool = True, **kwds: ArrayLike), so a
  ``**dict[str, ndarray]`` splat makes mypy match the first key against
  allow_pickle: bool. Documented ``# type: ignore[arg-type]`` on the four
  splat calls (acquire_mast_disruption_shots, build_mast_disruption_dataset,
  evaluate_mast_efm_neural_equilibrium, extract_mast_disruption_channels);
  the payloads never carry allow_pickle at runtime.

No runtime behaviour change; float() on an already-float rel-error is a
no-op. Pinned pre-commit mypy over src/ + validation/ passes; 86 module
tests (current-drive, sol-two-point, disruption-channel-recipes,
runaway-electron) pass.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)
Seat: c8f7

+++ validation/validate_current_drive.py
@@ -103,7 +103,7 @@ def critical_energy_rel_error(*, te_kev: float = 10.0, a_beam: float = 2.0, a_io
     """Relative error of the Stix critical energy against its closed form."""
     measured = float(nbi_critical_energy(te_kev, A_beam=a_beam, A_ion=a_ion))
     analytic = _E_CRIT_COEFFICIENT * te_kev * (a_beam / a_ion) ** (2.0 / 3.0)
-    return abs(measured - analytic) / analytic
+    return float(abs(measured - analytic) / analytic)
 
 
 def eccd_efficiency_rel_error(
